Feat: LSB & user menu (0826)
