+++ mc_web_console_new_front/src/app/Layouts/index.ts
@@ -1,0 +1,3 @@
+export * from './layoutHeader';
+export * from './mainLayout';
+export * from './verticalPageLayout';
