+++ mc_web_console_new_front/src/app/Layouts/layoutHeader/index.ts
@@ -1,0 +1,3 @@
+import LayoutHeader from './ui/LayoutHeader.vue';
+
+export { LayoutHeader };
