+++ mc_web_console_new_front/src/app/Layouts/mainLayout/index.ts
@@ -1,0 +1,3 @@
+import MainLayout from './ui/MainLayout.vue';
+
+export { MainLayout };
