+++ mc_web_console_new_front/src/app/Layouts/verticalPageLayout/index.ts
@@ -1,0 +1,3 @@
+import VerticalPageLayout from './VerticalPageLayout.vue';
+
+export { VerticalPageLayout };
