+++ mc_web_console_new_front/src/app/i18n/index.ts
