--- mc_web_console_new_front/src/app/language/en.json
@@ -1,22 +1,0 @@
-{
-  "MENU": {
-    "MANAGE": "Manage",
-    "MANAGE_DASHBOARD": "Dashboard",
-    "MANAGE_PROJECT": "Project",
-    "MANAGE_WORKLOADS": "Workloads",
-    "MANAGE_SERVICES_MANAGEMENT": "Services Management",
-    "MANAGE_CLOUD_META_DATA": "Cloud Meta Data",
-    "MANAGE_WORKFLOWS": "Workflows",
-    "ANALYTICS": "Analytics",
-    "ANALYTICS_MONITORINGS": "Monitorings",
-    "ENVIRONMENT": "Environment",
-    "ENVIRONMENT_CLOUD_SPS": "Cloud SPs",
-    "ACCOUNT_AND_ACCESS": "Account & Access",
-    "ACCOUNT_AND_ACCESS_ORGANIZATIONS": "Organizations",
-    "ACCOUNT_AND_ACCESS_CONTROL": "Access Control",
-    "ACCOUNT_AND_ACCESS_WORKSPACE_SETTINGS": "Workspace Settings"
-  },
-  "WORKSPACE": {
-    "WORKSPACE": "Workspace"
-  }
-}
