+++ mc_web_console_new_front/src/features/LSB/LSB/index.ts
@@ -1,0 +1,3 @@
+import LSB from './ui/LSB.vue';
+
+export { LSB };
