+++ mc_web_console_new_front/src/features/LSB/LSBMenuItem/index.ts
@@ -1,0 +1,3 @@
+import LSBMenuItem from './ui/LSBMenuItem.vue';
+
+export { LSBMenuItem };
