+++ mc_web_console_new_front/src/features/LSB/LSBRouterMenuItem/index.ts
@@ -1,0 +1,3 @@
+import LSBRouterMenuItem from './ui/LSBRouterMenuItem.vue';
+
+export { LSBRouterMenuItem };
