+++ mc_web_console_new_front/src/features/topbar/topbarNotifications/index.ts
@@ -1,0 +1,4 @@
+import TopBarNotifications from './ui/TopBarNotifications.vue';
+import TopBarNotificationContextMenu from '../../../widgets/topbarNotificationContextMenu/ui/TopBarNotificationContextMenu.vue';
+
+export { TopBarNotifications, TopBarNotificationContextMenu };
