+++ mc_web_console_new_front/src/features/topbar/topbarToolset/index.ts
@@ -1,0 +1,3 @@
+import TopBarToolset from './ui/TopBarToolset.vue';
+
+export { TopBarToolset };
