+++ mc_web_console_new_front/src/features/topbar/topbarWorkspaces/index.ts
@@ -1,0 +1,3 @@
+import TopBarWorkspaces from './ui/TopBarWorkspaces.vue';
+
+export { TopBarWorkspaces };
