+++ mc_web_console_new_front/src/main.ts
@@ -8,6 +8,7 @@ import './app/style/style.pcss';
 import { createPinia, PiniaVuePlugin } from 'pinia';
 import VueRouter from 'vue-router';
 import { McmpRouter } from '@/app/providers/router';
+import { i18n } from '@/app/i18n';
 
 const pinia = createPinia();
 Vue.use(PiniaVuePlugin);
@@ -16,6 +17,7 @@ Vue.use(VueRouter);
 
 new Vue({
   pinia,
+  i18n,
   router: McmpRouter.getRouter(),
   render: h => h(App),
 }).$mount('#app');
