+++ mc_web_console_new_front/src/pages/dashboard/vpc/VpcLSB.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import LSB from '@/widgets/subSidebar/LSB.vue';
+import LSB from '@/features/LSB/LSB/ui/LSB.vue';
 import { useMenuPerUserStore } from '@/entities/user/store/menuPerUserStore';
 import { storeToRefs } from 'pinia';
 import { PTextInput } from '@cloudforet-test/mirinae';
@@ -7,22 +7,33 @@ import { computed } from 'vue';
 
 const menuPerUserStore = useMenuPerUserStore();
 
+interface Props {
+  parentMenuName: string;
+  submenus: any[];
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  parentMenuName: '',
+  submenus: () => [],
+});
 const tempMenuSet = computed(() => {
-  return [
-    {
-      type: 'item',
-      label: 'All Projects',
-      icon: 'ic_dots-4-square',
-      to: '/dashboard/vpc',
-      hideFavorite: true,
-    },
-  ];
+  return props.submenus;
+  // return [
+  //   {
+  //     type: 'item',
+  //     label: 'All Projects',
+  //     icon: 'ic_dots-4-square',
+  //     to: '/dashboard/vpc',
+  //     hideFavorite: true,
+  //   },
+  // ];
 });
 </script>
 
 <template>
   <l-s-b class="vpc-l-s-b" :menu-set="tempMenuSet">
-    {{ tempMenuSet }}
+    <p-text-input class="vpc-search" placeholder="Search VPC" />
+
     <template #collapsible-contents-project>
       <p-text-input class="vpc-search" placeholder="Search VPC" />
     </template>
