+++ mc_web_console_new_front/src/shared/libs/store/lsb-store.ts
@@ -1,0 +1,19 @@
+import { defineStore } from 'pinia';
+
+export const useLSBStore = defineStore('lsb', {
+  state: () => ({
+    submenuInfo: {
+      parentMenuName: '',
+      submenus: [],
+    },
+  }),
+  getters: {},
+  actions: {
+    setSubmenuInfo(parentMenuName: string, submenus: any) {
+      this.submenuInfo = {
+        parentMenuName,
+        submenus: submenus,
+      };
+    },
+  },
+});
