+++ mc_web_console_new_front/src/widgets/gnbNavigationRail/index.ts
@@ -1,0 +1,3 @@
+import GNBNavigationRail from './ui/GNBNavigationRail.vue';
+
+export { GNBNavigationRail };
