+++ mc_web_console_new_front/src/widgets/gnbToolbox/index.ts
@@ -1,0 +1,3 @@
+import GNBToolbox from './ui/GNBToolbox.vue';
+
+export { GNBToolbox };
