+++ mc_web_console_new_front/src/widgets/gnbToolbox/ui/GNBToolbox.vue
