--- mc_web_console_new_front/src/widgets/layoutHeader/LayoutHeader.vue
@@ -1,66 +1,0 @@
-<script setup lang="ts">
-import TopBarWorkspaces from './ui/topBarWorkspaces/TopBarWorkspaces.vue';
-import TopBarToolset from './ui/topBarToolset/TopBarToolset.vue';
-import { reactive } from 'vue';
-
-const state = reactive({
-  openedMenu: '',
-});
-
-const hideMenu = () => {
-  state.openedMenu = '';
-};
-
-const handleOpenMenu = (menuId: any) => {
-  if (state.openedMenu === menuId) {
-    hideMenu();
-  } else {
-    state.openedMenu = menuId;
-  }
-};
-</script>
-
-<template>
-  <div class="top-bar">
-    <!-- TODO: 1. topbar workspaces -->
-    <top-bar-workspaces ref="topBarWorkspaceRef" class="top-bar-workspace" />
-    <!-- TODO: 2. topbar toolset -->
-    <top-bar-toolset
-      ref="topBarToolsetRef"
-      class="toolset"
-      :opened-menu="state.openedMenu"
-      @open-menu="handleOpenMenu"
-      @hide-menu="hideMenu"
-    />
-  </div>
-</template>
-
-<style lang="postcss" scoped>
-.top-bar {
-  @apply bg-white items-center justify-between border-b border-gray-200;
-  display: flex !important;
-  width: 100%;
-
-  &.admin-top-bar {
-    @apply bg-violet-900;
-    background: linear-gradient(90deg, #2c0f66 14.69%, #7d5dd2 100%);
-    box-shadow: 0 0.125rem 0.25rem 0 rgba(0, 0, 0, 0.12);
-  }
-
-  .top-bar-workspace {
-    flex: 0 0;
-  }
-
-  .middle-section {
-    @apply flex justify-center;
-    flex: 1 1;
-  }
-
-  .toolset {
-    @apply inline-flex items-center;
-    right: 0;
-    padding-right: 1.5rem;
-    flex: 0 0;
-  }
-}
-</style>
