+++ mc_web_console_new_front/src/widgets/menuCategory/index.ts
@@ -1,0 +1,3 @@
+import MenuCategory from './ui/MenuCategory.vue';
+
+export { MenuCategory };
