+++ mc_web_console_new_front/src/widgets/menuCategory/ui/sidebar-config.ts
