+++ mc_web_console_new_front/src/widgets/topbarNotificationContextMenu/index.ts
@@ -1,0 +1,3 @@
+import TopBarNotificationContextMenu from './ui/TopBarNotificationContextMenu.vue';
+
+export { TopBarNotificationContextMenu };
